design,foooter,about

+++ src/pages/ManageCampers.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState, useContext } from 'react';
-import {Modal, Button } from 'react-bootstrap';
+import {Modal, Button,Container } from 'react-bootstrap';
 import { addCamper, getCampers, updateCamper, deleteCampers, addManyCampers } from '../common/Database';
 import ManagebleTable from '../common/ManagebleTable';
 import { LanguageContext } from '../common/LanguageContext';
@@ -111,101 +111,27 @@ export default function ManageCampers() {
   const [newCampers, setNewCampers] = useState([]);
   
   const campers_columns = [
-    {
-      title: strings.camper_id[language],
-      field: 'id',
-      type: 'text',
+    {title: strings.camper_id[language], field: 'camper_id', type: 'text', invisible: true},
+    {title: strings.first_name[language], field: 'firstName', type: 'text'},
+    {title: strings.last_name[language], field: 'lastName', type: 'text'},
+    {title: strings.birth_date[language], field: 'birthDate', type: 'date', invisible: true},
+    {title: strings.gender[language], field: 'gender', type: 'select', options: 
+      {'male': strings.male[language], 'female': strings.female[language], 'other': strings.other[language] }
     },
-    {
-      title: strings.first_name[language],
-      field: 'firstName',
-      type: 'text'
+    {title: strings.city[language],field: 'city',type: 'text', invisible: true},
+    {title: strings.address[language],field: 'address',type: 'text', invisible: true},
+    {title: strings.frame[language], field: 'frame', type: 'text' },
+    { title: strings.disability_definition[language], field: 'disabilityDefinition', type: 'text' },
+    { title: strings.functioning_level[language], field: 'functioningLevel', type: 'select',
+      options: { 'high': strings.high[language], 'medium': strings.medium[language], 'low': strings.low[language] }
     },
-    {
-      title: strings.last_name[language],
-      field: 'lastName',
-      type: 'text'
-    },
-    {
-      title: strings.birth_date[language],
-      field: 'birthDate',
-      type: 'date'
-    },
-    {
-      title: strings.gender[language],
-      field: 'gender',
-      type: 'select',
-      options: {
-        'male': strings.male[language],
-        'female': strings.female[language],
-        'other': strings.other[language]
-      }
-    },
-    {
-      title: strings.city[language],
-      field: 'city',
-      type: 'text'
-    },
-    {
-      title: strings.address[language],
-      field: 'address',
-      type: 'text'
-    },
-    {
-      title: strings.frame[language],
-      field: 'frame',
-      type: 'text'
-    },
-    {
-      title: strings.disability_definition[language],
-      field: 'disabilityDefinition',
-      type: 'text'
-    },
-    {
-      title: strings.functioning_level[language],
-      field: 'functioningLevel',
-      type: 'select',
-      options: {
-        'high': strings.high[language],
-        'medium': strings.medium[language],
-        'low': strings.low[language]
-      }
-    },
-    {
-      title: strings.allergies[language],
-      field: 'allergies',
-      type: 'text'
-    },
-    {
-      title: strings.parent_name[language],
-      field: 'parentName',
-      type: 'text'
-    },
-    {
-      title: strings.parent_phone[language],
-      field: 'parentPhone',
-      type: 'text'
-    },
-    {
-      title: strings.branch[language],
-      field: 'branch',
-      type: 'text'
-    },
-    {
-      title: strings.tutor[language],
-      field: 'tutor',
-      type: 'text'
-    },
-    {
-      title: strings.tutor_phone[language],
-      field: 'tutorPhone',
-      type: 'text'
-    },
-    {
-      title: strings.special_comment[language],
-      field: 'comments',
-      type: 'textarea'
-    }
+    { title: strings.allergies[language], field: 'allergies', type: 'text' },
+    { title: strings.parent_name[language], field: 'parentName', type: 'text', invisible: true },
+    { title: strings.parent_phone[language], field: 'parentPhone', type: 'text', invisible: true },
+    { title: strings.branch[language], field: 'branch', type: 'text' },
+    { title: strings.tutor[language], field: 'tutor', type: 'text' },
+    { title: strings.tutor_phone[language], field: 'tutorPhone', type: 'text' },
+    { title: strings.special_comment[language], field: 'comments', type: 'textarea' }
   ]
 
   const globalActions = [
@@ -265,8 +191,9 @@ export default function ManageCampers() {
   }
 
   const handleDeleteCampers = (camperId) => {
-    
-    deleteCampers(camperId).then(getCampers).then(campersEscapeDate).then(campers => setCampers(campers));
+    const confirmed = window.confirm(strings.delete_confirm_camper[language]);
+    if (confirmed) 
+      deleteCampers(camperId).then(getCampers).then(campersEscapeDate).then(campers => setCampers(campers));
   }
 
   const handleAddManyCampers = (campers) => {
@@ -278,6 +205,7 @@ export default function ManageCampers() {
         if (!campers_columns.find((column) => column.field === key)) {
           delete camper[key];
         }
+        camper.birthDate = new Date(camper.birthDate);
       });
     });
     addManyCampers(newCampers).then(getCampers).then(campersEscapeDate).then(campers => setCampers(campers));
@@ -285,7 +213,7 @@ export default function ManageCampers() {
   }
 
   return (
-    <>
+    <Container>
       <CampersPreviewModal
         campers={newCampers}
         setCampers={setNewCampers}
@@ -304,7 +232,6 @@ export default function ManageCampers() {
         onDelete={handleDeleteCampers}
         editModal={true}
       />
-
-    </>
+    </Container>
   )
 }
